MISC: Refactor simple anisotropy runscript
